set width, height, update scroll bar / update styledcomponents setting
