+++ src/styles/globals.css
@@ -1,91 +1,91 @@
 :root {
   /* Main colors */
-  --primary: #6F6CF6; /* Same as purple500 */
-  --secondary: #E7ECFF; /* Same as purple100 */
-  --background: #FFFFFF;
-  --text: #010B13;
+  --primary: #6f6cf6; /* Same as purple500 */
+  --secondary: #e7ecff; /* Same as purple100 */
+  --background: #ffffff;
+  --text: #010b13;
 
   /* Grayscale */
-  --gray100: #F8F8F8;
-  --gray200: #DDE1E8;
-  --gray300: #B4BAC6;
+  --gray100: #f8f8f8;
+  --gray200: #dde1e8;
+  --gray300: #b4bac6;
   --gray400: #757880;
   --gray500: #363841;
 
   /* Purple scale */
-  --purple100: #E7ECFF;
-  --purple200: #BAC0F7;
-  --purple300: #8D8BFB;
-  --purple400: #6851F5;
-  --purple500: #6F6CF6;
+  --purple100: #e7ecff;
+  --purple200: #bac0f7;
+  --purple300: #8d8bfb;
+  --purple400: #6851f5;
+  --purple500: #6f6cf6;
 
   /* Additional colors */
-  --error: #FE4E39; /* Red */
-  --shadow: #5B6374; /* Shadow color */
+  --error: #fe4e39; /* Red */
+  --shadow: #5b6374; /* Shadow color */
 
   /* Font families */
-  --font-family-body: "SF Pro Text", Arial, sans-serif; /* Primary font for body text */
-  --font-family-heading: "SF Pro Text", Georgia, serif; /* Primary font for headings */
-  --font-family-monospace: "SF Mono", "Courier New", monospace; /* Font for code or monospace needs */
+  --font-family-body: 'SF Pro Text', Arial, sans-serif; /* Primary font for body text */
+  --font-family-heading: 'SF Pro Text', Georgia, serif; /* Primary font for headings */
+  --font-family-monospace: 'SF Mono', 'Courier New', monospace; /* Font for code or monospace needs */
 
   /* Font weights */
-  --font-weight-light: 300;     /* Light */
-  --font-weight-regular: 400;   /* Regular */
+  --font-weight-light: 300; /* Light */
+  --font-weight-regular: 400; /* Regular */
   --font-weight-semi-bold: 600; /* Semi-bold */
 
   /* Font sizes */
-  --font-size-xs: 0.75rem;    /* 12px */
-  --font-size-sm: 0.875rem;   /* 14px */
-  --font-size-md: 1rem;       /* 16px */
-  --font-size-lg: 1.25rem;    /* 20px */
-  --font-size-xl: 1.375rem;   /* 22px */
-  --font-size-xxl: 1.5rem;    /* 24px */
+  --font-size-xs: 0.75rem; /* 12px */
+  --font-size-sm: 0.875rem; /* 14px */
+  --font-size-md: 1rem; /* 16px */
+  --font-size-lg: 1.25rem; /* 20px */
+  --font-size-xl: 1.375rem; /* 22px */
+  --font-size-xxl: 1.5rem; /* 24px */
 
   /* Border radii */
   --border-radius-default: 0.6rem; /* 9.6px */
-  --border-radius-md: 0.5rem;      /* 8px */
-  --border-radius-lg: 1rem;        /* 16px */
+  --border-radius-md: 0.5rem; /* 8px */
+  --border-radius-lg: 1rem; /* 16px */
 
   /* Spacing */
-  --spacing-xs: 0.5rem;   /* 8px */
-  --spacing-sm: 1rem;     /* 16px */
-  --spacing-md: 1.5rem;   /* 24px */
-  --spacing-lg: 2rem;     /* 32px */
-  --spacing-xl: 3rem;     /* 48px */
+  --spacing-xs: 0.5rem; /* 8px */
+  --spacing-sm: 1rem; /* 16px */
+  --spacing-md: 1.5rem; /* 24px */
+  --spacing-lg: 2rem; /* 32px */
+  --spacing-xl: 3rem; /* 48px */
 
   /* Breakpoints */
-  --breakpoint-mobile-min: 479px;  /* min-width for mobile (portrait) */
-  --breakpoint-mobile-max: 767px;  /* max-width for mobile (landscape) */
-  --breakpoint-tablet-min: 768px;  /* min-width for tablets (portrait) */
+  --breakpoint-mobile-min: 479px; /* min-width for mobile (portrait) */
+  --breakpoint-mobile-max: 767px; /* max-width for mobile (landscape) */
+  --breakpoint-tablet-min: 768px; /* min-width for tablets (portrait) */
   --breakpoint-tablet-max: 1023px; /* max-width for tablets (landscape) */
-  --breakpoint-desktop: 1024px;    /* min-width for desktops */
+  --breakpoint-desktop: 1024px; /* min-width for desktops */
 }
 
 @media (prefers-color-scheme: dark) {
   :root {
     /* Main colors */
-    --primary: #8967F6; /* Same as purple500 */
-    --secondary: #2B215A; /* Same as purple100 */
-    --background: #010B13; /* Same as gray100 */
-    --text: #F8F8F8;
+    --primary: #8967f6; /* Same as purple500 */
+    --secondary: #2b215a; /* Same as purple100 */
+    --background: #010b13; /* Same as gray100 */
+    --text: #f8f8f8;
 
     /* Grayscale */
-    --gray100: #010B13;
+    --gray100: #010b13;
     --gray200: #363841;
     --gray300: #757880;
-    --gray400: #B4BAC6;
-    --gray500: #DDE1E8;
+    --gray400: #b4bac6;
+    --gray500: #dde1e8;
 
     /* Purple scale */
-    --purple100: #2B215A;
-    --purple200: #44328C;
-    --purple300: #5C43B5;
-    --purple400: #7252E0;
-    --purple500: #8967F6;
+    --purple100: #2b215a;
+    --purple200: #44328c;
+    --purple300: #5c43b5;
+    --purple400: #7252e0;
+    --purple500: #8967f6;
 
     /* Additional colors */
-    --error: #FF6A62; /* Red */
-    --shadow: #5B6374; /* Shadow color */
+    --error: #ff6a62; /* Red */
+    --shadow: #5b6374; /* Shadow color */
   }
 
   html {
@@ -149,16 +149,56 @@ input:disabled {
 
 html,
 body {
+  height: 100%;
   max-width: 100vw;
   overflow-x: hidden;
 }
 
 body {
   color: var(--text);
-  background-color: var(--gray100);
   font-family: var(--font-family-body);
-  width: 360px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 100vh;
+}
+
+#layout-wrapper {
+  width: 100%;
+  max-width: 360px;
   height: 815px;
+  display: flex;
+  flex-direction: column;
+  overflow-y: auto;
+  overflow-x: hidden;
+}
+
+#main-content {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  background-color: var(--gray100);
+  width: 100%;
+  height: 100%;
+}
+
+/* 스크롤바 스타일 (Webkit 기반 브라우저) */
+::-webkit-scrollbar {
+  width: 10px; /* 스크롤바의 너비 */
+  height: 12px; /* 스크롤바의 높이 (수평 스크롤바) */
+}
+
+::-webkit-scrollbar-track {
+  background: var(--gray200); /* 트랙 배경 색상 */
+}
+
+::-webkit-scrollbar-thumb {
+  background: var(--gray300); /* 핸들의 배경 색상 */
+  border-radius: var(--border-radius-default); /* 모서리 둥글게 */
+}
+
+::-webkit-scrollbar-thumb:hover {
+  background: var(--gray400); /* 핸들 마우스 오버 시 색상 */
 }
 
 a {
@@ -167,8 +207,8 @@ a {
 }
 
 /* Input Styles */
-input[type="text"],
-input[type="password"],
+input[type='text'],
+input[type='password'],
 textarea {
   width: 100%;
   padding: var(--spacing-sm); /* 1rem */
@@ -183,20 +223,20 @@ textarea {
   transition: border-color 0.2s;
 }
 
-input[type="text"]:focus,
-input[type="password"]:focus,
+input[type='text']:focus,
+input[type='password']:focus,
 textarea:focus {
   border-color: var(--primary);
 }
 
-input[type="text"]::placeholder,
-input[type="password"]::placeholder,
+input[type='text']::placeholder,
+input[type='password']::placeholder,
 textarea::placeholder {
   color: var(--gray400);
 }
 
-input[type="text"]:disabled,
-input[type="password"]:disabled,
+input[type='text']:disabled,
+input[type='password']:disabled,
 textarea:disabled {
   background-color: var(--gray100);
   color: var(--gray300);
